Search: move SearchDebouncer to its own file and drop inline comments
