+++ Amperfy/Screens/ViewController/AlbumsCommonVCInteractions.swift
@@ -554,8 +554,6 @@ class AlbumsCommonVCInteractions {
       reloadListViewCB?()
       updateContentUnavailable()
     }
-    // Typing: debounce the expensive contains[cd] fetch. Cleared text: restore
-    // the full list immediately (and drop any pending search).
     if searchText.isEmpty {
       searchDebouncer.runImmediately(performSearch)
     } else {
