+++ Amperfy/Screens/ViewController/AlbumsVC.swift
@@ -307,8 +307,6 @@ class AlbumsVC: SingleSnapshotFetchedResultsTableViewController<AlbumMO> {
   }
 
   override func updateSearchResults(for searchController: UISearchController) {
-    // The list refresh now runs inside common's debounced search (via reloadListViewCB),
-    // so it no longer fires (and faults the A-Z section index) on every keystroke.
     common.updateSearchResults(for: self.searchController)
   }
 }
