+++ Amperfy/Screens/ViewController/ArtistsVC.swift
@@ -399,8 +399,6 @@ class ArtistsVC: SingleSnapshotFetchedResultsTableViewController<ArtistMO> {
       }
       updateContentUnavailable()
     }
-    // Typing: debounce the expensive contains[cd] fetch. Cleared text: restore
-    // the full list immediately (and drop any pending search).
     if searchText.isEmpty {
       searchDebouncer.runImmediately(performSearch)
     } else {
