+++ Amperfy/Screens/ViewController/PlaylistAdd/PlaylistAddAlbumsVC.swift
@@ -160,8 +160,6 @@ class PlaylistAddAlbumsVC: SingleSnapshotFetchedResultsTableViewController<Album
   }
 
   override func updateSearchResults(for searchController: UISearchController) {
-    // The list refresh now runs inside common's debounced search (via reloadListViewCB),
-    // so it no longer fires on every keystroke.
     common.updateSearchResults(for: self.searchController)
   }
 }
