+++ Amperfy/Screens/ViewController/PlaylistAdd/PlaylistAddArtistsVC.swift
@@ -204,8 +204,6 @@ class PlaylistAddArtistsVC: SingleSnapshotFetchedResultsTableViewController<Arti
         }}
       }
     }
-    // Typing: debounce the expensive contains[cd] fetch. Cleared text: restore
-    // the full list immediately (and drop any pending search).
     if searchText.isEmpty {
       searchDebouncer.runImmediately(performSearch)
     } else {
