+++ Amperfy/Screens/ViewController/PlaylistAdd/PlaylistAddSongsVC.swift
@@ -258,8 +258,6 @@ class PlaylistAddSongsVC: SingleFetchedResultsTableViewController<SongMO>, Playl
       }
       tableView.reloadData()
     }
-    // Typing: debounce the expensive contains[cd] fetch. Cleared text: restore
-    // the full list immediately (and drop any pending search).
     if searchText.isEmpty {
       searchDebouncer.runImmediately(performSearch)
     } else {
